Move vulcand logic to each module
