+++ receiver/vulcand/vulcand.go
@@ -1,9 +1,6 @@
 package vulcand
 
 import (
-	"bytes"
-	"encoding/json"
-	"fmt"
 	"strings"
 
 	"github.com/dtan4/paus-gitreceive/receiver/model"
@@ -12,11 +9,7 @@ import (
 )
 
 const (
-	VulcandKeyBase = "/vulcand"
-)
-
-var (
-	httpBackendJSON string
+	vulcandKeyBase = "/vulcand"
 )
 
 func RegisterInformation(etcd *store.Etcd, application *model.Application, baseDomain string, webContainer *model.Container) ([]string, error) {
@@ -40,66 +33,4 @@ func RegisterInformation(etcd *store.Etcd, application *model.Application, baseD
 	}
 
 	return identifiers, nil
-}
-
-// {"Type": "http"}
-func setBackend(etcd *store.Etcd, application *model.Application, baseDomain string) error {
-	key := fmt.Sprintf("%s/backends/%s/backend", VulcandKeyBase, application.ProjectName)
-
-	if err := etcd.Set(key, httpBackendJSON); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// {"Type": "http", "BackendId": "$identifier", "Route": "Host(`$identifier.$base_domain`) && PathRegexp(`/`)", "Settings": {"TrustForwardHeader": true}}
-func setFrontend(etcd *store.Etcd, application *model.Application, identifier, baseDomain string) error {
-	key := fmt.Sprintf("%s/frontends/%s/frontend", VulcandKeyBase, identifier)
-	frontend := Frontend{
-		Type:      "http",
-		BackendId: application.ProjectName,
-		Route:     fmt.Sprintf("Host(`%s.%s`) && PathRegexp(`/`)", strings.ToLower(identifier), strings.ToLower(baseDomain)),
-		Settings: FrontendSettings{
-			TrustForwardHeader: true,
-		},
-	}
-
-	b, err := json.Marshal(frontend)
-
-	if err != nil {
-		return errors.Wrap(err, "Failed to generate vulcand frontend JSON.")
-	}
-
-	// json.Marshal generates HTML-escaped JSON string
-	b = bytes.Replace(b, []byte("\\u0026"), []byte("&"), -1)
-	json := string(b)
-
-	if err := etcd.Set(key, json); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// {"URL": "http://$web_container_host_ip:$web_container_port"}
-func setServer(etcd *store.Etcd, application *model.Application, container *model.Container, baseDomain string) error {
-	key := fmt.Sprintf("%s/backends/%s/servers/%s", VulcandKeyBase, application.ProjectName, container.ContainerId)
-	server := Server{
-		URL: fmt.Sprintf("http://%s:%s", container.HostIP(), container.HostPort()),
-	}
-
-	b, err := json.Marshal(server)
-
-	if err != nil {
-		return errors.Wrap(err, "Failed to generate vulcand server JSON.")
-	}
-
-	json := string(b)
-
-	if err := etcd.Set(key, json); err != nil {
-		return err
-	}
-
-	return nil
 }
